+++ variables.tf
@@ -23,6 +23,12 @@ variable "name" {
   description = "(Required) The name of the cluster."
 }
 
+variable "rbac_security_identity_group" {
+  description = "(Optional) The name of the RBAC security identity group for use with Google security groups in Kubernetes RBAC. Group name must be in format `gke-security-groups@yourdomain.com`."
+  type        = string
+  default     = null
+}
+
 variable "min_master_version" {
   type        = string
   description = "(Required) The Kubernetes minimal version of the masters. If set to 'latest' it will pull latest available version in the selected region."
@@ -184,4 +190,34 @@ variable "enable_binary_authorization" {
   type        = bool
   description = "(Optional) Enable BinAuthZ Admission controller"
   default     = false
+}
+
+# ----------------------------------------------------------------------------------------------------------------------
+# MODULE CONFIGURATION PARAMETERS
+# These variables are used to configure the module.
+# ----------------------------------------------------------------------------------------------------------------------
+
+variable "module_enabled" {
+  type        = bool
+  description = "(Optional) Whether to create resources within the module or not."
+  default     = true
+}
+
+variable "module_timeouts" {
+  description = "(Optional) How long certain operations (per resource type) are allowed to take before being considered to have failed."
+  type        = any
+  # type = object({
+  #   google_container_cluster = optional(object({
+  #     create = optional(string)
+  #     update = optional(string)
+  #     delete = optional(string)
+  #   }))
+  # })
+  default = {}
+}
+
+variable "module_depends_on" {
+  type        = any
+  description = "(Optional) A list of external resources the module depends_on."
+  default     = []
 }